// Backend/src/types/Statistics.ts
export interface RawUserStats {
  rank: string | null;
  address: string;
  totalTicketsWithBonus: string;
  totalWins: string;
  totalTickets: string;
  maxTickets: string;
  consecutiveRounds: string;
  contractAddress: string;
  token: string | null;
  roundId: string;
  totalElements?: string;
}